Finished part 2
Had the right idea, but a stupid mistake made me lose a few hours. I
was using the index of the shape and jet instead of the modulo of the
indices... Thus never finding any cycle.

In the iterim I switched to a List<byte> and a bitmask solution to store
the grid data. Not really needed as the existing sparse grid was enough
once cycle calculations were brought in.