Day 22 much better performance
Used a pseudo-hash as key instead of a tuple with the 4 last price
changes)

The pseudo hash is just each price change on 5 bits.

This makes computing the next hash easy, keep the first 15 bits of the previous hash, shift left
5 and OR the next price change on 5 bits.

Also use PSeq for the final computation. Not a huge gain, probably due to contention on the
ConcurrentDictionary.