Day 23 done
graph problems require graph algorithms so I went over to the wikipedia
page for graph algos and quickly found that it's the maximum
clique problem. Went over to the known algos section for that problem and selected Bron Kerbosh
because it looked simple.

Implementation went without a hitch. Lost a bit of time because the
solution is comma separated and not just the strings joined ... duh.